feat!: Move arrays from prelude into new extension
