+++ hugr-core/src/std_extensions/collections.rs
@@ -1,3 +1,4 @@
 //! List type and operations.
 
+pub mod array;
 pub mod list;
