+++ hugr-llvm/src/extension/collections.rs
@@ -1,3 +1,4 @@
 //! Emission logic for collections.
 
+pub mod array;
 pub mod list;
